+++ src/Integration/Aws/SimpleS3/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Added
+
+- `download()`, `has()` and `upload()` methods.
+
 ## 0.1.0
 
 First version